refactor : Game : 게임 시작과 종료를 알리는 문구 출력
