+++ src/main/java/softeer2nd/Main.java
@@ -1,9 +1,10 @@
 package softeer2nd;
 
+import softeer2nd.manager.Game;
+
 public class Main {
     public static void main(String[] args) {
         Game game = new Game();
-
         game.run();
     }
 }
