+++ src/main/java/softeer2nd/manager/Game.java
@@ -1,15 +1,24 @@
-package softeer2nd;
+package softeer2nd.manager;
 
 import softeer2nd.chess.Board;
 
 import java.util.Scanner;
 
 public class Game {
+    private Board board;
+
+    public Game() {
+        this.board = new Board();
+    }
+
     public void run() {
-        Board board = new Board();
+        System.out.println("game start!");
 
         while (true){
             board.initialize();
+
+            System.out.println("input : 'start' | 'end'");
+
             Scanner in = new Scanner(System.in);
             String input = in.next();
 
@@ -18,6 +27,8 @@ public class Game {
 
             System.out.println(board.showBoard());
         }
+
+        System.out.println("game end!");
     }
 
     private boolean verifyInput(String input){
@@ -25,7 +36,7 @@ public class Game {
             return true;
         else{
             if(!input.equals("end"))
-                System.out.println("Input Error!");
+                System.out.println("Input Error!"); // TODO : Exception 생성 (Business Exception => InvalidInputException이 적당할듯)
             return false;
         }
     }
