+++ src/main/java/softeer2nd/manager/Input.java
@@ -1,0 +1,4 @@
+package softeer2nd.manager;
+
+public class Input {
+}
